Working on ReceiptCoreService Unit Tests.

+++ WebRifa.Blazor.Core/CoreServices/ReceiptCoreService.cs
@@ -6,11 +6,17 @@ public class ReceiptCoreService(IReceiptRepository receiptRepository, ITicketRep
     private readonly IReceiptRepository _receiptRepository = receiptRepository;
     private readonly ITicketRepository _ticketRepository = ticketRepository;
 
-    public async Task DeleteReceipt(ReceiptDeleteCommand command, CancellationToken cancellation)
+    public async Task DeleteReceiptAsync(ReceiptDeleteCommand command, CancellationToken cancellation)
     {
         Receipt receipt = await _receiptRepository.GetAsync(command.ReceiptId, cancellation);
+        if (receipt == null) {
+            throw new NullReferenceException();
+        }
 
         await _receiptRepository.DeleteAsync(receipt, cancellation);
-        await _ticketRepository.DeleteRangeAsync(receipt.Tickets, cancellation);
+
+        if (receipt.Tickets is not null) {
+            await _ticketRepository.DeleteRangeAsync(receipt.Tickets, cancellation);
+        }
     }
 }
